ctests: add ertroundtrip lable and invert a lyx22x
In many cases, round trip with older formats involves exporting ERT
or preamble code in the backwards conversion. In the forwards
conversion, if this code is not parsed, often errors can result.
However, in many cases, especially for older formats, it might not
be worth the time or code complexity to address these cases. Such
tests are labled "ertroundtrip".

This commit also inverts a currently failing lyx22x test under the
label "ertroundtrip" since the above paragraph is my best guess as
to why that test is failing. It is likely not worth the time to fix
it, especially since the APA7 layout wasn't even shipped for LyX
2.2.x.