DocBook: implement theorems.
This is a minimal implementation, as DocBook lacks a serious way of encoding all of this. Maybe a <formalpara> could do the trick, but I'd need to find a way to shoehorn a title through the styles (i.e. a first complete tag):

Theorem: Bla bla

<formalpara>
<title>Theorem</title>
<para>Bla bla</para>
</formalpara>

This would also only be a solution for single-paragraph things, as formalpara only allows one paragraph. Or a sidebar, but it's semantically very remote.