Address failures of Seminar examples on new TL20
Compilation of our Seminar example file fails on updated TL20. The
maintainer of "Seminar" is not planning to fix the core issue and
states the following (in a private email with permission to quote):

  it is a problem with the new hook management of the current latex.ltx
  seminar is a quite old package and there is no reason to use it with a
  new LaTeX format. It won't be fixed, so the usual way is to use the
  package latexrealease to get the old hook management.

This commit adds a note to the example files explaining the
workaround of exporting to a .tex file and prepending the following
line:

  \RequirePackage[2020-02-02]{latexrelease}

We now invert the relevant tests.