Make all exectuable python scripts use python3
This is only relevant on linux/unix if running the scripts from a shell.
These two were the last where the call still used an unversioned python.

This has no reflex on the way that lyx calls the scripts or the python
version used since the #! "shebang line" is ignored.