Powerdot examples: work around undefined ref
The hook that defines the "lastslide" label is interfered with for
some reason on an updated TeX Live 2020. We thus avoid referencing
the label and instead show only the current slide number on each
slide. We leave instructions in the preamble for how to get back the
previous behavior if desired.