ctests: invert Hebrew tutorial LuaTeX tests
With an updated TeX Live 2020 an assertion is given. The assertion
was reported on the LuaTeX ML [1], and will be converted into the
following error as of LuaTeX svn commit r7385:

  ! the word doesn't start with a character

The error could likely be worked around with a change to the
document but since the document compiles without error with XeTeX I
did not change anything since I do not understand the issue.

[1]
https://www.mail-archive.com/search?l=mid&q=20201030030430.4y4i6bdpwx633qs7%40vbox-VirtualBox