Move "LyX Documentation Tools" toolbar to "Others"
This move is done because not many users likely use this toolbar.

Those who use it now need to go through the extra submenu "Other
Toolbars" and also need to move the toolbar from the bottom to the
top (if they want previous behavior). But once that is done, the
toolbar stays across sessions.