Fix correction of row in getPitAndRowNearY
Now that the ascent of the paragraph metrics is not necessarily the
ascent of the first row (see 907f0207), the test is too strict. Now it
is more symmetrical wrt the !up case.

Note that changing this part of code is associated with a
non-negligible risk of creating a bug elsewhere.

Fixes part of bug #12123.